refactor(native): use light expo browser polyfill, unify hooks
