+++ package.json
@@ -34,8 +34,6 @@
     "eslint-plugin-import": "^2.25.4",
     "eslint-plugin-react": "^7.28.0",
     "eslint-plugin-react-hooks": "^4.3.0",
-    "expo-asset": "^8.4.6",
-    "expo-file-system": "^13.1.2",
     "expo-gl": "^11.1.1",
     "jest": "^27.4.7",
     "ogl": "^0.0.81",
@@ -49,16 +47,13 @@
     "typescript": "^4.5.5"
   },
   "dependencies": {
-    "base64-arraybuffer": "^1.0.1",
-    "react-merge-refs": "^1.1.0",
+    "@expo/browser-polyfill": "^1.0.0",
     "react-reconciler": "^0.26.2",
     "react-use-measure": "^2.1.1",
     "suspend-react": "^0.0.8"
   },
   "peerDependencies": {
     "expo": ">=43.0",
-    "expo-asset": ">=8.4",
-    "expo-file-system": ">=13.0",
     "expo-gl": ">=11.0",
     "ogl": ">=0.0.72",
     "react": ">=17.0",
@@ -76,12 +71,6 @@
       "optional": true
     },
     "expo-gl": {
-      "optional": true
-    },
-    "expo-asset": {
-      "optional": true
-    },
-    "expo-file-system": {
       "optional": true
     }
   },
