+++ src/native/index.ts
@@ -1,4 +1,5 @@
+import '@expo/browser-polyfill'
+
 export * from './Canvas'
 export * from './events'
-export * from './hooks'
 export * from '../'
