+++ src/web/index.ts
@@ -1,4 +1,3 @@
 export * from './Canvas'
 export * from './events'
-export * from './hooks'
 export * from '../'
